// src/types/certificate.ts
export interface CertificateData {
  certificateDate: string;
  pin: string;
  taxpayerName: string;
  emailAddress: string;
  lrNumber: string;
  building: string;
  streetRoad: string;
  cityTown: string;
  county: string;
  district: string;
  taxArea: string;
  station: string;
  poBox: string;
  postalCode: string;
  taxObligation: string;
  effectiveFromDate: string;
  effectiveTill: string;
  status: string;
}